existing resources checks all namespace, instead of jst default it left blank
