+++ pkg/engine/engine.go
@@ -3,140 +3,17 @@ package engine
 import (
 	jsonpatch "github.com/evanphx/json-patch"
 	"github.com/golang/glog"
-	"github.com/minio/minio/pkg/wildcard"
 	types "github.com/nirmata/kyverno/pkg/apis/policy/v1alpha1"
 	client "github.com/nirmata/kyverno/pkg/dclient"
 	"github.com/nirmata/kyverno/pkg/info"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	v1helper "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 )
-
-func ListResourcesThatApplyToPolicy(client *client.Client, policy *types.Policy) map[string]resourceInfo {
-	// key uid
-	resourceMap := map[string]resourceInfo{}
-	for _, rule := range policy.Spec.Rules {
-		// Match
-		for _, k := range rule.MatchResources.Kinds {
-			// kind -> resource
-			gvr := client.DiscoveryClient.GetGVRFromKind(k)
-			// Namespace
-			namespace := "default"
-			if rule.MatchResources.Namespace != nil {
-				namespace = *rule.MatchResources.Namespace
-			}
-			if k == "Namespace" {
-				namespace = ""
-			}
-			// Check if exclude namespace is not clashing
-			if rule.ExcludeResources.Namespace != nil && *rule.ExcludeResources.Namespace == namespace {
-				// as the namespace is excluded
-				continue
-			}
-
-			// List resources
-			list, err := client.ListResource(k, namespace, rule.MatchResources.Selector)
-			if err != nil {
-				glog.Errorf("unable to list resource for %s with label selector %s", gvr.Resource, rule.MatchResources.Selector.String())
-				glog.Errorf("unable to apply policy %s rule %s. err: %s", policy.Name, rule.Name, err)
-				continue
-			}
-			var selector labels.Selector
-			// exclude label selector
-			if rule.ExcludeResources.Selector != nil {
-				selector, err = v1helper.LabelSelectorAsSelector(rule.ExcludeResources.Selector)
-				if err != nil {
-					glog.Error(err)
-				}
-			}
-			for _, res := range list.Items {
-				// exclude label selectors
-				if selector != nil {
-					set := labels.Set(res.GetLabels())
-					if selector.Matches(set) {
-						// if matches
-						continue
-					}
-				}
-				var name *string
-				// match
-				// name
-				// wild card matching
-				name = rule.MatchResources.Name
-				if name != nil {
-					// if does not match then we skip
-					if !wildcard.Match(*name, res.GetName()) {
-						continue
-					}
-				}
-				// exclude
-				// name
-				// wild card matching
-				name = rule.ExcludeResources.Name
-				if name != nil {
-					// if matches then we skip
-					if wildcard.Match(*name, res.GetName()) {
-						continue
-					}
-				}
-				gvk := res.GroupVersionKind()
-
-				ri := resourceInfo{Resource: res, Gvk: &metav1.GroupVersionKind{Group: gvk.Group,
-					Version: gvk.Version,
-					Kind:    gvk.Kind}}
-
-				resourceMap[string(res.GetUID())] = ri
-
-			}
-		}
-	}
-	return resourceMap
-}
 
 // ProcessExisting checks for mutation and validation violations of existing resources
 func ProcessExisting(client *client.Client, policy *types.Policy) []*info.PolicyInfo {
 	glog.Infof("Applying policy %s on existing resources", policy.Name)
 	// key uid
 	resourceMap := ListResourcesThatApplyToPolicy(client, policy)
-
-	// for _, rule := range policy.Spec.Rules {
-	// 	for _, k := range rule.Kinds {
-	// 		// kind -> resource
-	// 		gvr := client.DiscoveryClient.GetGVRFromKind(k)
-	// 		// label selectors
-	// 		// namespace ? should it be default or allow policy to specify it
-	// 		namespace := "default"
-	// 		if rule.ResourceDescription.Namespace != nil {
-	// 			namespace = *rule.ResourceDescription.Namespace
-	// 		}
-	// 		if k == "Namespace" {
-	// 			namespace = ""
-	// 		}
-
-	// 		list, err := client.ListResource(k, namespace, rule.ResourceDescription.Selector)
-	// 		if err != nil {
-	// 			glog.Errorf("unable to list resource for %s with label selector %s", gvr.Resource, rule.Selector.String())
-	// 			glog.Errorf("unable to apply policy %s rule %s. err: %s", policy.Name, rule.Name, err)
-	// 			continue
-	// 		}
-	// 		for _, res := range list.Items {
-	// 			name := rule.ResourceDescription.Name
-	// 			gvk := res.GroupVersionKind()
-	// 			if name != nil {
-	// 				// wild card matching
-	// 				if !wildcard.Match(*name, res.GetName()) {
-	// 					continue
-	// 				}
-	// 			}
-	// 			ri := resourceInfo{resource: res, gvk: &metav1.GroupVersionKind{Group: gvk.Group,
-	// 				Version: gvk.Version,
-	// 				Kind:    gvk.Kind}}
-
-	// 			resourceMap[string(res.GetUID())] = ri
-
-	// 		}
-	// 	}
-	// }
 	policyInfos := []*info.PolicyInfo{}
 	// for the filtered resource apply policy
 	for _, v := range resourceMap {
